requirements.txt
Flask
pdfminer.six
spacy